requirements.txt
behave
allure-behave
allure-pytest
pytest
requests
selenium
appium-python-client
webdriver-manager
pytest-xdist
python-dotenv
pytest-cov
loguru
